Add files via upload
A very basic pre-processing tactic I thought of for separating image of notes into "lines". for now it shows the graph of the sum of grayscale values (squared) where local minima are the lines where text is most probably written and local maxima are white lines. the idea is to split the images along the local maxima.